# server/openapi.json
{
  "components": {
    "schemas": {
...
      "ErrorBody": {
        "properties": {
          "error": {
            "$ref": "#/components/schemas/ErrorCode"
          }
        },
        "required": [
          "error"
        ],
        "title": "ErrorBody",
        "type": "object"
      },
      "ErrorCode": {
        "enum": [
          "database_unavailable",
          "unsupported_protocol",
          "registration_closed",
          "registration_needs_email",
          "registration_invite_required",
          "email_taken",
          "invalid_email",
          "invalid_credentials",
          "account_disabled",
          "unauthorized",
          "device_pending",
          "device_revoked",
          "device_not_found",
          "invalid_token",
          "token_expired",
          "recovery_failed",
          "rate_limited",
          "invalid_request",
          "email_unavailable",
          "email_not_verified",
          "setup_not_needed",
          "invalid_setup_token",
          "invalid_totp",
          "csrf_failed",
          "setting_locked",
          "not_found"
        ],
        "title": "ErrorCode",
        "type": "string"
      },
...
    }
  },
  "info": {
    "description": "The Tildeck sync server API.",
    "title": "Tildeck",
    "version": "protocol-1"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/api/devices/{device_id}/revoke": {
      "post": {
        "description": "Revokes a device (the current one signs out). Its token stops working\nimmediately.",
        "operationId": "revokeDevice",
        "parameters": [
          {
            "in": "path",
            "name": "device_id",
            "required": true,
            "schema": {
              "title": "Device Id",
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "Tildeck-Protocol",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "integer"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Tildeck-Protocol"
            }
          },
          {
            "in": "header",
            "name": "authorization",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Successful Response"
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            },
            "description": "Not signed in"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            },
            "description": "No such device"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            },
            "description": "Invalid request"
          }
        },
        "summary": "Revoke",
        "tags": [
          "account"
        ]
      }
    },
...
  }
}